Add mittigations for Page Reload after sending
When the page is reloaded after the successfull sending of a message,
you're prompted to confirm that you want to re-submit your posted data.
And if you do so, another message is sent. By redirecting the user back
to the same page, this is avoided.

As this was the intended, toggleable, behaviour of debounce, the config
parameter was removed.